requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==4.0.2
django-cors-headers==3.11.0
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
libgravatar==1.0.0
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
django-widget-tweaks==1.4.9
Faker==8.10.3
numpy==1.22.2
scikit-surprise==1.1.1
pyspark==3.2.1
pandas==1.4.1
django-autocomplete-light==3.9.1
coverage==6.3.2
django-profanity-filter==0.2.1
gunicorn==20.1.0
django-heroku
